Fix DevTools window not showing on Windows (#303)...
Add "Quick build instructions for Windows" section indocs/Build-instructions.

Fix invalid types for window handles on Windows 64-bit (#302).

Also fix long long for other pointers such as in PaintBuffer.
Also fix long long for uint types.
Get rid of invalid/unnecessary long conversions eg. from int64 type.

+++ src/client_handler/render_handler.cpp
@@ -94,7 +94,7 @@ bool RenderHandler::StartDragging(CefRefPtr<CefBrowser> browser,
 {
     REQUIRE_UI_THREAD();
     return RenderHandler_StartDragging(browser, drag_data,
-                                       static_cast<long>(allowed_ops), x, y);
+                                       allowed_ops, x, y);
 }
 
 
@@ -102,5 +102,5 @@ void RenderHandler::UpdateDragCursor(CefRefPtr<CefBrowser> browser,
                                      DragOperation operation)
 {
     REQUIRE_UI_THREAD();
-    RenderHandler_UpdateDragCursor(browser, static_cast<long>(operation));
+    RenderHandler_UpdateDragCursor(browser, operation);
 }
